requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pypdf==3.17.1
google-generativeai==0.3.1
scikit-learn==1.3.2
python-docx==1.1.0
python-dotenv==1.0.0